Start splitting driver
